requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
distro==1.9.0
ecdsa==0.19.1
fastapi==0.121.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.10.0
openai==2.7.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
tqdm==4.67.1
types-pyasn1==0.6.0.20250914
types-python-jose==3.5.0.20250531
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
